# Tracy-based expr evaluation profiler