--- a/.gitmodules
+++ b/.gitmodules
@@ -1,3 +1,0 @@
-[submodule "src/libexpr/tracy"]
-	path = src/libexpr/tracy
-	url = git@github.com:wolfpld/tracy.git