--- a/src/libexpr/eval.cc
+++ b/src/libexpr/eval.cc
@@ -55,6 +55,11 @@
 
 #endif
 
+#define TRACY_TRACE(evalstate, expr) \
+    std::ostringstream tracyss; \
+    tracyss << evalstate.positions[(expr)->getPos()] << " " << (expr)->showExprType(); \
+    ZoneTransientN(nix, tracyss.str().c_str(), true);
+
 using json = nlohmann::json;
 
 namespace nix {
@@ -1204,29 +1209,34 @@ void Expr::eval(EvalState & state, Env & env, Value & v)
 
 void ExprInt::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v = this->v;
 }
 
 
 void ExprFloat::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v = this->v;
 }
 
 void ExprString::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v = this->v;
 }
 
 
 void ExprPath::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v = this->v;
 }
 
 
 void ExprAttrs::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v.mkAttrs(state.buildBindings(attrs.size() + dynamicAttrs.size()).finish());
     auto dynamicEnv = &env;
 
@@ -1311,6 +1321,7 @@ void ExprAttrs::eval(EvalState & state, Env & env, Value & v)
 
 void ExprLet::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     /* Create a new environment that contains the attributes in this
        `let'. */
     Env & env2(state.allocEnv(attrs->attrs.size()));
@@ -1329,6 +1340,7 @@ void ExprLet::eval(EvalState & state, Env & env, Value & v)
 
 void ExprList::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     state.mkList(v, elems.size());
     for (auto [n, v2] : enumerate(v.listItems()))
         const_cast<Value * &>(v2) = elems[n]->maybeThunk(state, env);
@@ -1337,6 +1349,7 @@ void ExprList::eval(EvalState & state, Env & env, Value & v)
 
 Value * ExprList::maybeThunk(EvalState & state, Env & env)
 {
+    TRACY_TRACE(state, this);
     if (elems.empty()) {
         return &state.vEmptyList;
     }
@@ -1346,6 +1359,7 @@ Value * ExprList::maybeThunk(EvalState & state, Env & env)
 
 void ExprVar::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     Value * v2 = state.lookupVar(&env, *this, false);
     state.forceValue(*v2, pos);
     v = *v2;
@@ -1373,7 +1387,7 @@ static std::string showAttrPath(EvalState & state, Env & env, const AttrPath & a
 
 void ExprSelect::eval(EvalState & state, Env & env, Value & v)
 {
-    ZoneScopedN("select");
+    TRACY_TRACE(state, this)
     Value vTmp;
     PosIdx pos2;
     Value * vAttrs = &vTmp;
@@ -1438,6 +1452,7 @@ void ExprSelect::eval(EvalState & state, Env & env, Value & v)
 
 void ExprOpHasAttr::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this)
     Value vTmp;
     Value * vAttrs = &vTmp;
 
@@ -1463,6 +1478,7 @@ void ExprOpHasAttr::eval(EvalState & state, Env & env, Value & v)
 
 void ExprLambda::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this)
     v.mkLambda(&env, this);
 }
 
@@ -1720,6 +1736,7 @@ void EvalState::callFunction(Value & fun, size_t nrArgs, Value * * args, Value &
 
 void ExprCall::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     Value vFun;
     fun->eval(state, env, vFun);
 
@@ -1797,6 +1814,7 @@ https://nixos.org/manual/nix/stable/language/constructs.html#functions.)", symbo
 
 void ExprWith::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     Env & env2(state.allocEnv(1));
     env2.up = &env;
     env2.values[0] = attrs->maybeThunk(state, env);
@@ -1807,6 +1825,7 @@ void ExprWith::eval(EvalState & state, Env & env, Value & v)
 
 void ExprIf::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     // We cheat in the parser, and pass the position of the condition as the position of the if itself.
     (state.evalBool(env, cond, pos, "while evaluating a branch condition") ? then : else_)->eval(state, env, v);
 }
@@ -1814,6 +1833,7 @@ void ExprIf::eval(EvalState & state, Env & env, Value & v)
 
 void ExprAssert::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     if (!state.evalBool(env, cond, pos, "in the condition of the assert statement")) {
         std::ostringstream out;
         cond->show(state.symbols, out);
@@ -1825,12 +1845,14 @@ void ExprAssert::eval(EvalState & state, Env & env, Value & v)
 
 void ExprOpNot::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v.mkBool(!state.evalBool(env, e, getPos(), "in the argument of the not operator")); // XXX: FIXME: !
 }
 
 
 void ExprOpEq::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     Value v1; e1->eval(state, env, v1);
     Value v2; e2->eval(state, env, v2);
     v.mkBool(state.eqValues(v1, v2, pos, "while testing two values for equality"));
@@ -1839,6 +1861,7 @@ void ExprOpEq::eval(EvalState & state, Env & env, Value & v)
 
 void ExprOpNEq::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     Value v1; e1->eval(state, env, v1);
     Value v2; e2->eval(state, env, v2);
     v.mkBool(!state.eqValues(v1, v2, pos, "while testing two values for inequality"));
@@ -1847,6 +1870,7 @@ void ExprOpNEq::eval(EvalState & state, Env & env, Value & v)
 
 void ExprOpAnd::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v.mkBool(state.evalBool(env, e1, pos, "in the left operand of the AND (&&) operator") && state.evalBool(env, e2, pos, "in the right operand of the AND (&&) operator"));
 }
 
@@ -1859,12 +1883,14 @@ void ExprOpOr::eval(EvalState & state, Env & env, Value & v)
 
 void ExprOpImpl::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     v.mkBool(!state.evalBool(env, e1, pos, "in the left operand of the IMPL (->) operator") || state.evalBool(env, e2, pos, "in the right operand of the IMPL (->) operator"));
 }
 
 
 void ExprOpUpdate::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     Value v1, v2;
     state.evalAttrs(env, e1, v1, pos, "in the left operand of the update (//) operator");
     state.evalAttrs(env, e2, v2, pos, "in the right operand of the update (//) operator");
@@ -1903,6 +1929,7 @@ void ExprOpUpdate::eval(EvalState & state, Env & env, Value & v)
 
 void ExprOpConcatLists::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     Value v1; e1->eval(state, env, v1);
     Value v2; e2->eval(state, env, v2);
     Value * lists[2] = { &v1, &v2 };
@@ -1941,6 +1968,7 @@ void EvalState::concatLists(Value & v, size_t nrLists, Value * * lists, const Po
 
 void ExprConcatStrings::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     NixStringContext context;
     std::vector<BackedStringView> s;
     size_t sSize = 0;
@@ -2033,12 +2061,14 @@ void ExprConcatStrings::eval(EvalState & state, Env & env, Value & v)
 
 void ExprPos::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     state.mkPos(v, pos);
 }
 
 
 void ExprBlackHole::eval(EvalState & state, Env & env, Value & v)
 {
+    TRACY_TRACE(state, this);
     state.error("infinite recursion encountered")
         .debugThrow<InfiniteRecursionError>();
 }