--- a/src/libexpr/local.mk
+++ b/src/libexpr/local.mk
@@ -21,6 +21,12 @@ libexpr_LDFLAGS += -lboost_context $(THREAD_LDFLAGS)
 ifdef HOST_LINUX
   libexpr_LDFLAGS += -ldl
 endif
+ifneq ($(TRACY_PROFILER), no)
+libexpr_LDFLAGS += -ltracy
+# We have to set TRACY_ENABLE to have tracy actually send the trace
+# events, it's no-op without them.
+libexpr_CXXFLAGS += -DTRACY_ENABLE=1
+endif
 
 # The dependency on libgc must be propagated (i.e. meaning that
 # programs/libraries that use libexpr must explicitly pass -lgc),