This is a WIP profiler for the Nix code itself, not the cpp Nix evaluator. The goal is to provide some tooling to improve the Nixpkgs evaluation performance.

This is a WIP Nix profiler using [Tracy](https://github.com/wolfpld/tracy). The idea is to trace the evaluation of the expression atoms to figure out what is expensive. We are roughly tracing the thunk evaluation. Nix is a lazy language, I think this is what makes most sense.

[Short demo video](https://social.alternativebit.fr/media/4fee7e0b64e3d7fd05102e2d3b6789099ceb506c6cd9fe4398efd5aea2e13779.nix-tracy-demo-short)

## TODO before Undrafting

This is not ready yet. We need to fix at least those two issues before undrafting this PR.

**HELP WELCOME** :)

### Elephant in the room: tracy SEGFAULTS!!!!!!!!!

We have those weird "???" frames in Tracy. Hovering them with the mouse segfault the viewer.

Removing the line and the column in the position pretty printer makes them go away. Sadly, removing the column and position from the trace also makes the profiler useless :D

There's something very wrong in our client code, but I couldn't figure out what. For now. This is the first item on my TODO list.

### Profiler performance

There's room for improvement. Random datapoint, with the profiler enabled:
```
[nix-shell:~/code-root/github.com/NixOS/nix]$ time ./outputs/out/bin/nix-instantiate -A nixosTests.pleroma ../nixpkgs
warning: you did not specify '--add-root'; the result might be removed by the garbage collector
/nix/store/z6nqb4zkx59hqn00rpzpzq4fcp34y9pa-vm-test-run-pleroma.drv

real    1m41,835s
user    2m51,728s
sys     0m11,388s
```

Without the profiler:
```
[nix-shell:~/code-root/github.com/NixOS/nix]$ time nix-instantiate -A nixosTests.pleroma ../nixpkgs
warning: you did not specify '--add-root'; the result might be removed by the garbage collector
/nix/store/z6nqb4zkx59hqn00rpzpzq4fcp34y9pa-vm-test-run-pleroma.drv

real    0m12,478s
user    0m9,998s
sys     0m1,431s
```

That's not a surprise, I did not make any effort here, each profiler macro is generating **at least** 3 allocations.

We should run perf on this and see what's wrong.


### Move out the trace macros from the eval call sites

See https://github.com/NixOS/nix/pull/9967#issuecomment-1933695607

### Is this useful/refine design?

We should spend some time using this profiler. Can we spot issues with it? Can we remove some useless noise from the traces?

There's no magic trick here, we need to spend time using the tool to improve it.

## How to test?

1. `nix build github:picnoir/nix/pic/tracy-profiler`
2. Open tracy. `nix-shell -p tracy`. Then, either lauch the graphical profiler to watch the capture in real time (`tracy`), either record the run through the headless capture utility (`capture -o capture.tracy`).
3. Evaluate something with the `nix-instantiate` utility you built in step 2. EG. `./result/bin/nix-instantiate -A nixosTests.pleroma path-to/nixpkgs/`
4. Enjoy the [operation game](https://en.wikipedia.org/wiki/Operation_(game)), avoid the `???` with your mouse to prevent segfaults. Scroll-wheel to zoom/unzoom, left click to move the trace.


# Priorities and Process

Add :+1: to [pull requests you find important](https://github.com/NixOS/nix/pulls?q=is%3Aopen+sort%3Areactions-%2B1-desc).

The Nix maintainer team uses a [GitHub project board](https://github.com/orgs/NixOS/projects/19) to [schedule and track reviews](https://github.com/NixOS/nix/tree/master/maintainers#project-board-protocol).