@nixos-discourse 2024-03-12
This pull request has been mentioned on **NixOS Discourse**. There might be relevant details there:

https://discourse.nixos.org/t/nix-flamegraph-or-profiling-tool/33333/14