@Atemu 2024-06-29
The tracy patch does not appear to fix the segfault for me, it still crashes when I dare to move my cursor over its window.

I've opened a trace of a NixOS eval and am on a relatively recent nixos-unstable.